Use gettimeofday() on Mac OS X.
Mac OS X lacks CLOCK_MONOTONIC as well as clock_gettime and
pthread_condattr_setclock.

There is some concern that this might not be safe on other
platforms that also happen to lack CLOCK_MONOTONIC or that
there might be better alternatives, so we only do this on
Mac OS X for now.