Initialize epbase item.
epbase nn_list_item struct was not initialized, leading to failures when
an item was removed from a list before it was added to the list.

Now nn_epbase_init() initialize the item to nill item.

This fixes the "Conditional jump or move depends on uninitialised
value(s)" seen in valgrind in the tcp test.

Signed-off-by: Martin Sustrik <sustrik@250bpm.com>