Fixed windows signalling bug (fixes #189)
Assuming that there may be a lag between nn_efd_signal and the data
available in nn_efd_unsignal, we set NN_SOCK_FLAG_IN/OUT if we have a
sucessful return from nn_efd_wait. Setting this flag is safe at any
time. And I expect that it should not create a bottleneck