fixes #647 Initial cmake support for non-Windows platforms
This change is the first step towards a unified CMake based build for
all nanomsg platforms.  With this it is possible to build nanomsg
using cmake on at least Linux, MacOS X, illumos, and Windows systems.
However, there remains work to be done, particularly with respect to
delivery of documentation and the nanocat binary.

The eventual goal will be to remove support for the autotools, and
settle on a single cmake based toolchain.

This work is the result of a collaborative effort by several authors:
Jack R. Dunaway (@JackDunaway), Garrett D'Amore (@gdamore),
Franklin Mathieu (@Snaipe) and Shiva (@shiva).