Sets pipebase's data-pointer to NULL during initialization.
A freshly initialized pipe(base) does not have any additional data
assigned to it. However, a receiver of a pipe might not necessarily have
the knowledge, whether or not someone else might have set the pointer to
point to additional data or not. Since the pipe is not set to NULL
during initialization of itself, the pointer is likely to contain an
arbitrary value and thus may cause later checks by the receivers of the
pipe to fail. Therefore, the data-pointer should be initialized with
value NULL in nn_pipebase_init to prevent undefined behaviour or even
crashes.