Adding examples html file
This was created by Tim Dysinger. The original eventually ended up offline for several months or more. This has been pulled from archive.org and modified to condense it into a single file. It has gone through light HTML edits and the removal of unnecessary Javascript (all of it) and CSS (most of it). Images were converted to base64 to embed directly into this file. 

No changes were made to any of the visible text, images, or example code. Links were only modified to remove the reference to archive.org links.