Updates to Getting Started.
This set of revisions separates the C from the asciidoc, and breaks
each pattern into separate documents.  A number of errors were fixed,
and the example code was hardened a bit to make it more robust and
correct for newbies who might be expecting it to Just Work.

This also sets the stage for us to make versions of this for NNG, mangos,
and other languages.