fixes #978 nanomsg IPC on Windows is very fragile
This is critical for hardening use cases with other IPC clients
such as nng or mangos.  The old code made some very incorrect
assumptions about the atomicity of named pipes and ReadFile.

We've changed the code so that if ReadFile (or WSARecv incidentally)
ever returns a partial read, we keep going.

This solves a critial assertion error, and greatly improves
the crash resistance of nanomsg when using IPC on Windows.